requirements.txt
flask==3.0.3
MarkUpSafe==2.1.5
blinker==1.8.2
click==8.1.7
colorama==0.4.6
itsdangerous==2.2.0
jinja2==3.1.4
werkzeug==3.0.3
kafka-python==2.0.2
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.3.4
python-dotenv==1.0.1
gevent==24.2.1